# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Student Information System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all users (admin only)
      parameters:
        - name: role
          in: query
          schema:
            type: string
        - name: departmentId
          in: query
          schema:
            type: ["integer", "null"]
        - name: page
          in: query
          schema:
            type: integer
        - name: pageSize
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserListResponse"
    post:
      operationId: createUser
      tags: [users]
      summary: Create a new user (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserInput"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{userId}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get user by ID
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    patch:
      operationId: updateUser
      tags: [users]
      summary: Update a user
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: User updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    delete:
      operationId: deleteUser
      tags: [users]
      summary: Delete a user
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
  /drives:
    get:
      operationId: listDrives
      tags: [drives]
      summary: List placement drives
      parameters:
        - name: departmentId
          in: query
          schema:
            type: ["integer", "null"]
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of drives
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Drive"
    post:
      operationId: createDrive
      tags: [drives]
      summary: Create a placement drive
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DriveInput"
      responses:
        "201":
          description: Drive created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Drive"
  /drives/{driveId}:
    get:
      operationId: getDrive
      tags: [drives]
      summary: Get drive by ID
      parameters:
        - name: driveId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Drive details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Drive"
    patch:
      operationId: updateDrive
      tags: [drives]
      summary: Update a placement drive
      parameters:
        - name: driveId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DriveUpdate"
      responses:
        "200":
          description: Drive updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Drive"
    delete:
      operationId: deleteDrive
      tags: [drives]
      summary: Delete a placement drive
      parameters:
        - name: driveId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
...
    User:
      type: object
      required: [id, email, role, fullName]
      properties:
        id:
          type: integer
        email:
          type: string
        role:
          type: string
          enum: [SUPER_ADMIN, CENTRAL_PLACEMENT, DEPT_PLACEMENT, SRC, STUDENT]
        fullName:
          type: string
        departmentId:
          type: ["integer", "null"]
        departmentName:
          type: ["string", "null"]
        studentId:
          type: ["integer", "null"]
        createdAt:
          type: string
          format: date-time
    UserInput:
      type: object
      required: [email, password, role, fullName]
      properties:
        email:
          type: string
        password:
          type: string
        role:
          type: string
        fullName:
          type: string
        departmentId:
          type: ["integer", "null"]
    UserUpdate:
      type: object
      properties:
        email:
          type: string
        fullName:
          type: string
        role:
          type: string
        departmentId:
          type: ["integer", "null"]
        password:
          type: ["string", "null"]
    UserListResponse:
      type: object
      required: [data, total, page, pageSize]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/User"
        total:
          type: integer
        page:
          type: integer
        pageSize:
          type: integer
...
    Drive:
      type: object
      required: [id, title, company, status]
      properties:
        id:
          type: integer
        title:
          type: string
        company:
          type: string
        description:
          type: ["string", "null"]
        eligibleDepartments:
          type: array
          items:
            type: string
        minCgpa:
          type: ["number", "null"]
        minTenthPercent:
          type: ["number", "null"]
        minTwelfthPercent:
          type: ["number", "null"]
        maxBacklogs:
          type: ["integer", "null"]
        driveDate:
          type: ["string", "null"]
          format: date
        applicationDeadline:
          type: ["string", "null"]
          format: date
        package:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        jobType:
          type: ["string", "null"]
        status:
          type: string
          enum: [UPCOMING, ONGOING, COMPLETED, CANCELLED]
        applicantCount:
          type: integer
        selectedCount:
          type: integer
        createdAt:
          type: string
          format: date-time
    DriveInput:
      type: object
      required: [title, company]
      properties:
        title:
          type: string
        company:
          type: string
        description:
          type: ["string", "null"]
        eligibleDepartments:
          type: array
          items:
            type: string
        minCgpa:
          type: ["number", "null"]
        minTenthPercent:
          type: ["number", "null"]
        minTwelfthPercent:
          type: ["number", "null"]
        maxBacklogs:
          type: ["integer", "null"]
        driveDate:
          type: ["string", "null"]
        applicationDeadline:
          type: ["string", "null"]
        package:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        jobType:
          type: ["string", "null"]
        status:
          type: string
    DriveUpdate:
      type: object
      properties:
        title:
          type: string
        company:
          type: string
        description:
          type: ["string", "null"]
        eligibleDepartments:
          type: array
          items:
            type: string
        minCgpa:
          type: ["number", "null"]
        driveDate:
          type: ["string", "null"]
        applicationDeadline:
          type: ["string", "null"]
        package:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        status:
          type: string
...